Setting up automation scripts

+++ Load-Generator-Analyser/main-load-profiler.sh
@@ -1,7 +1,16 @@
 #!/bin/bash
 
-rm collected-profiles/*.json
+# if the collected-profiles directory doesn't exist, create it
+if [ ! -d "collected-profiles" ]; then
+    mkdir collected-profiles
+fi
 
+# if there's any previous results, remove them
+if ls collected-profiles/*.json 1> /dev/null 2>&1; then
+    rm collected-profiles/*.json
+fi
+
+# triggering the load generator and the analyser
 for NumReplicas in 5 4 3 2 1; do
     for RequestInterval in 50 40 30 20 10; do
         echo "========================================================="
